refactor(probe): migrate udp planning into per-protocol plugin
